+++ rest_framework/serializers/__init__.py
@@ -5,7 +5,9 @@ Serializers.
 from .serializers import Serializer, ListSerializer
 from .fields import (
     BooleanField, CharField, IntegerField, FloatField, ListField,
-    TimeField, DateField, DateTimeField
+    TimeField, DateField, DateTimeField,
+    JsonField, DictField,
+    SerializerMethodField,
 )
 from .exceptions import ValidationError
 
@@ -14,6 +16,8 @@ __all__ = (
     # fields
     'BooleanField', 'CharField', 'IntegerField', 'FloatField', 'ListField',
     'TimeField', 'DateField', 'DateTimeField',
+    'JsonField', 'DictField',
+    'SerializerMethodField',
 
     # serializers
     'Serializer', 'ListSerializer',