+++ rest_framework/utils/html.py
@@ -76,3 +76,32 @@ def parse_html_list(dictionary, prefix=''):
 
     # Return result.
     return [ret[item] for item in sorted(ret.keys())]
+
+
+def parse_html_dict(dictionary, prefix=''):
+    """
+    Used to support dictionary values in HTML forms.
+
+    {
+        'profile.username': 'example',
+        'profile.email': 'example@example.com',
+    }
+        -->
+    {
+        'profile': {
+            'username': 'example',
+            'email': 'example@example.com'
+        }
+    }
+    """
+    ret = MultiValueDict()
+    regex = re.compile(r'^%s\.(.+)$' % re.escape(prefix))
+    for field in dictionary:
+        match = regex.match(field)
+        if not match:
+            continue
+        key = match.groups()[0]
+        value = dictionary.getlist(field)
+        ret.setlist(key, value)
+
+    return ret